Add files via upload
As part of the ambitious “Made in China 2025” blueprint, Chinese officials have set the semiconductor industry a goal of reaching US$305 billion in output by 2030 and meeting 80% of domestic demand. Aiming high, the Chinese semiconductor industry seems frustrating as for its development despite the government’s support. As most commonly known, talents and capital resources are the fundamental drivers for the development of tech-intensive industry. Guided by this concern,  our research focuses on the two most important factors determining the prospect of this industry—intellectual training and financial investment. Our project aims to provide deep insights into the current bottlenecks and advise on future development to help achieve the long-term sustainable thriving of the Chinese semiconductor industry.